Phase 4: the 64-bit object memory
om_mt keeps the Blue Book's object-pointer encoding -- low bit tags a
SmallInteger, otherwise twice a table index -- so the guaranteed pointers
still have the values Chapter 27 gives them and the interpreter compiles
against either memory without a change.  What goes away is the 16-bit
ceiling: 63-bit SmallIntegers, a table that grows on demand, and object
bodies bounded only by the host allocator.

The object table is kept deliberately, and threading is the reason.
become: is a swap of two entries rather than a stop-the-world heap scan,
pinning for SDL and foreign code is free because the entry is the
identity, and the entry is the natural home for the lock word Phase 7
needs.  One indirection buys atomic identity mutation.

Also here: a versioned little-endian snapshot format that stores no
reference counts, since a collection rebuilds them from reachability on
load, and a marking collector carrying forward the CompiledMethod
literal-frame fix from Phase 3.

Two build defects fixed on the way.  The binary was written to one path
for both memories, so switching OM= left a newer st80 in place and make
reported nothing to do while silently running the other build.  And the
census histogram was sized by ST_HUGE_SIZE, which is 2^31 under the new
memory.

90 checks pass, clean under ASAN and TSAN.  The Blue Book suite skips
under OM=mt and the new suite skips under OM=bb, so both builds are
green.

+++ src/interp/trace.c
@@ -42,7 +42,7 @@ ST_print_object(st_oop p, char *buf, size_t buflen)
     if (buflen == 0)
         return;
     if (OM_is_int(p)) {
-        snprintf(buf, buflen, "%d", (int) OM_int_value(p));
+        snprintf(buf, buflen, "%lld", (long long) OM_int_value(p));
         return;
     }
     if (p == ST_NIL) {
